requirements.txt
pyrogram
tgcrypto
motor
pillow
python-dotenv
dnspython
